Fix for 3587919

+++ sveditor/plugins/net.sf.sveditor.core.tests/src/net/sf/sveditor/core/tests/content_assist/TestContentAssistSystem.java
@@ -53,10 +53,12 @@ public class TestContentAssistSystem extends TestCase {
 		
 		if (fProject != null) {
 			TestUtils.deleteProject(fProject);
+			fProject = null;
 		}
 		
-		if (fTmpDir.exists()) {
+		if (fTmpDir != null && fTmpDir.exists()) {
 			TestUtils.delete(fTmpDir);
+			fTmpDir = null;
 		}
 	}
